- code optimization for account validation functions done
- code optimization for account validation functions done
