+++ app/views/walletSetup/createAccount/index.js
@@ -12,6 +12,7 @@ import check from './check.svg';
 import user from './user.svg';
 import lock from './lock.svg';
 import { LOADER_COLOR } from '../../../constants/index';
+import { validateData } from '../../../general/validations/index';
 
 import * as CreateAccountAction from '../../../reducers/createAccount/action';
 
@@ -89,72 +90,9 @@ class CreateAccount extends Component {
     return disable;
   };
 
-  validateData = (event, value, name) => {
-    event.preventDefault();
-    let validationResult = '';
-    if (name === 'accountName') {
-      const regex = /^[a-zA-Z ]{2,30}$/;
-      const result = regex.test(value);
-      if (result) {
-        validationResult = { errorText: '' };
-      } else {
-        validationResult = { errorText: 'Enter a valid name' };
-      }
-    } else if (name === 'password') {
-      validationResult = this.validPass(value);
-    } else if (name === 'confirmPassword') {
-      validationResult = this.validRepass(value);
-    }
-    return validationResult;
-  };
-
-  validPass = value => {
-    const errorObj = {};
-    if (value === '') {
-      errorObj.errorText = "Password field can't be empty";
-    } else if (value.length < 8) {
-      errorObj.errorText =
-        'Make your password with 8 characters or more. It can be any combination of letters, numbers, and symbols.';
-    } else if (value.match(/[A-Z]/) === null) {
-      errorObj.errorText = 'Password field should contain one Capital letter';
-    } else if (value.match(/[0-9]/) === null) {
-      errorObj.errorText = 'Password field should contain one number';
-    } else {
-      errorObj.errorText = '';
-    }
-    return errorObj;
-  };
-
-  validEmail = value => {
-    const errorObj = {};
-    const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    if (value === '') {
-      errorObj.errorText = "Account Name field can't be empty";
-    } else if (re.test(String(value).toLowerCase())) {
-      errorObj.errorText = '';
-    } else {
-      errorObj.errorText = 'You need to specify a valid account name';
-    }
-    return errorObj;
-  };
-
-  validRepass = value => {
-    const { password } = this.state;
-    const errorObj = {};
-    if (value === '') {
-      errorObj.errorText = "Re-enter password field can't be empty";
-    } else if (value !== password) {
-      errorObj.errorText =
-        'Password and re-enter password fields must be the same.';
-    } else {
-      errorObj.errorText = '';
-    }
-    return errorObj;
-  };
-
   setAccountName(e) {
     const accountName = e.target.value;
-    const isValid = this.validateData(e, accountName, 'accountName');
+    const isValid = validateData(e, accountName, 'accountName', '');
     this.setState(
       {
         accountName,
@@ -168,7 +106,7 @@ class CreateAccount extends Component {
 
   setPassword(e) {
     const password = e.target.value.trim();
-    const isValid = this.validateData(e, password, 'password');
+    const isValid = validateData(e, password, 'password', '');
     this.setState(
       {
         password,
@@ -196,7 +134,12 @@ class CreateAccount extends Component {
 
   setConfirmPassword(e) {
     const confirmPassword = e.target.value.trim();
-    const isValid = this.validateData(e, confirmPassword, 'confirmPassword');
+    const isValid = validateData(
+      e,
+      confirmPassword,
+      'confirmPassword',
+      this.state.password
+    );
     this.setState(
       {
         confirmPassword,
